pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
python-dateutil = "^2.8.1"
pytest = "^6.1.1"
selenium = "^3.141.0"
pymongo = {extras = ["srv"], version = "^3.11.3"}
mongomock = "^3.22.1"
flask-Login = "^0.5.0"
oauthlib = "^3.1.0"
loggly-python-handler = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"